Fix cycle dep
