+++ src/status.js
@@ -1,17 +1,16 @@
-import { toDoList } from './index';
+const toDoList = JSON.parse(localStorage.getItem('toDoList'));
 export const status = (() => {
-
   const addCheckBoxHandlers = () => {
     [...document.querySelectorAll('.check')].forEach((box) => {
       box.addEventListener('change', () => {
         box.nextElementSibling.classList.toggle('checked');
-        box.chacked = !box.checked;
-        box.parentElement.dataset.id
         toDoList[box.parentElement.dataset.id].completed = box.checked;
         localStorage.setItem('toDoList', JSON.stringify(toDoList));
-      })
-    })
-  }
+      });
+    });
+
+
+  };
 
   return { addCheckBoxHandlers };
 })();
