+++ webpack.config.js
@@ -19,7 +19,7 @@ module.exports = {
       use: ['style-loader', 'css-loader'],
     }],
   },
-
+  devtool: 'inline-source-map',
   output: {
     filename: 'main.js',
     path: path.resolve(__dirname, 'dist'),
